modification des tests + ajouts

+++ TP1/prog1.py
@@ -2,8 +2,8 @@ import fonctions as f
 
 while True:
     try:
-        a = float(input("Entrez le premier nombre (base entière) : "))
-        b = float(input("Entrez le second nombre (exposant entier) : "))
+        a = int(input("Entrez le premier nombre (base entière) : "))
+        b = int(input("Entrez le second nombre (exposant entier) : "))
 
         res = f.puissance(a, b)
         print(f"{a} à la puissance {b} donne {res}\n")
